requirements.txt
absl-py==0.12.0
aiobotocore==2.1.2
aiohttp==3.8.6
aioitertools==0.8.0
aiosignal==1.2.0
alembic==1.7.7
altair==4.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.1
attrs==22.2.0
backports.zoneinfo==0.2.1
bleach==1.5.0
blinker==1.5
botocore==1.23.24
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
cloudpickle==2.2.1
comm==0.1.4
commonmark==0.9.1
coverage==6.2
cycler==0.11.0
databricks-cli==0.17.8
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
Django==3.2.25
django-cors-headers==3.10.1
djangorestframework==3.15.1
docker==5.0.3
docutils==0.18.1
flake8==5.0.4
Flask==2.0.3
Flask-Cors==5.0.0
flatbuffers==1.12
fpdf==1.7.2
frozenlist==1.2.0
fsspec==2022.1.0
gast==0.4.0
gcsfs==2022.1.0
gender-guesser==0.4.0
gitdb==4.0.9
GitPython==3.1.18
google-api-core==2.8.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.1
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
greenlet==2.0.2
grpcio==1.48.2
h5py==3.1.0
html5lib==0.9999999
idna==3.7
idna-ssl==1.1.0
imageio==2.15.0
imageio-ffmpeg==0.5.1
imbalanced-learn==0.8.1
imblearn==0.0
importlib-metadata==3.10.1
importlib-resources==5.4.0
iniconfig==1.1.1
instaloader==4.9.5
ipywidgets==7.8.5
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.1
jsonschema==3.2.0
jupyterlab-pygments==0.1.2
jupyterlab_widgets==1.1.11
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==23.4.1
kiwisolver==1.3.1
Mako==1.1.6
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.7.0
memoized-property==1.0.3
mistune==0.8.4
mlflow==1.23.1
moviepy==1.0.3
multidict==5.2.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nltk==3.6.7
nnfs==0.5.1
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.1
patsy==1.0.1k
Pillow==8.4.0
pkginfo==1.10.0
plotly==5.18.0
pluggy==1.0.0
proglog==0.1.10
prometheus-client==0.17.1
prometheus_flask_exporter==0.23.1
promise==2.3
protobuf==3.19.6
py==1.11.0
pyarrow==6.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pycparser==2.21
pydeck==0.6.2
pydub==0.25.1
pyflakes==2.5.0
PyJWT==2.4.0
Pympler==1.1
pyparsing==3.1.4
PyPDF2==3.0.1
pyrsistent==0.18.0
pytest==7.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pywin32==227
pywin32-ctypes==0.2.3
pywinpty==1.1.6
PyYAML==6.0.1
querystring-parser==1.2.4
readme-renderer==34.0
regex==2023.8.8
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==12.0.1
rsa==4.9
s3fs==2022.1.0
scikeras==0.4.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.3
six==1.15.0
smmap==5.0.0
SQLAlchemy==1.4.54
sqlparse==0.4.4
statsmodels==0.12.2
streamlit==1.10.0
tabulate==0.8.10
tenacity==8.2.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-addons==0.14.0
tensorflow-datasets==4.5.2
tensorflow-estimator==2.6.0
tensorflow-hub==0.16.0
tensorflow-metadata==1.2.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
textblob==0.17.1
threadpoolctl==3.1.0
toml==0.10.2
tomli==1.2.3
toolz==0.12.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.64.1
tweepy==4.6.0
twine==3.8.0
typeguard==2.13.3
typing-extensions==3.7.4.3
tzdata==2024.2
tzlocal==4.2
urllib3==1.26.19
validators==0.20.0
vit-keras==0.1.0
waitress==2.0.0
watchdog==2.3.1
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.10
wincertstore==0.2
wrapt==1.12.1
xgboost==1.5.2
yapf==0.32.0
yarl==1.7.2
zipp==3.6.0
